REF move version update code into functions
