+++ conda_forge_tick/update_recipe/__init__.py
@@ -1,0 +1,2 @@
+from .build_number import update_build_number, DEFAULT_BUILD_PATTERNS  # noqa
+from .version import update_version  # noqa
